pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pitlane-simgate"
version = "0.1.0"
description = "Sim-Gate Orchestrator: build scenarios from logs, run sweeps, evaluate policy gates, and attest promotion decisions."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "PitlaneOps" }]
license = { text = "Apache-2.0" }
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pyyaml>=6.0.1",
  "jinja2>=3.1.4",
  "pynacl>=1.5.0"
]

[project.scripts]
simgate = "pitlane_simgate.cli:app"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
